requirements.txt
requests
ipykernel
bs4
selenium==4.0.0a3
pyyaml==5.2